+++ tests/codegen/service-codegen-tests/src/main/kotlin/com/test/DefaultServiceGeneratorTest.kt
@@ -11,6 +11,11 @@ import java.nio.file.Path
 import java.nio.file.Paths
 
 internal fun main() {
+    generateServiceGeneratorTest()
+    generateServiceConstraintsTest()
+}
+
+internal fun generateServiceGeneratorTest() {
     val modelPath: Path = Paths.get("model", "service-generator-test.smithy")
     val defaultModel = ModelAssembler()
         .discoverModels()
@@ -50,7 +55,7 @@ internal fun main() {
             ObjectNode.builder().withMember("framework", Node.from("ktor")).build(),
         )
         .build()
-    val outputDir: Path = Paths.get("build", "generated-service").also { Files.createDirectories(it) }
+    val outputDir: Path = Paths.get("build", "service-generator-test").also { Files.createDirectories(it) }
     val manifest: FileManifest = FileManifest.create(outputDir)
 
     val context: PluginContext = PluginContext.builder()
@@ -101,7 +106,73 @@ internal fun main() {
     manifest.writeFile("src/main/kotlin/$packagePath/auth/Validation.kt", bearerValidation)
 
     val settingGradleKts = """
-        rootProject.name = "generated-project"
+        rootProject.name = "service-generator-test"
+        includeBuild("../../../../../")
+    """.trimIndent()
+    manifest.writeFile("settings.gradle.kts", settingGradleKts)
+}
+
+internal fun generateServiceConstraintsTest() {
+    val modelPath: Path = Paths.get("model", "service-constraints-test.smithy")
+    val defaultModel = ModelAssembler()
+        .discoverModels()
+        .addImport(modelPath)
+        .assemble()
+        .unwrap()
+    val serviceName = "ServiceConstraintsTest"
+    val packageName = "com.test"
+
+    val packagePath = packageName.replace('.', '/')
+
+    val settings: ObjectNode = ObjectNode.builder()
+        .withMember("service", Node.from("$packageName#$serviceName"))
+        .withMember(
+            "package",
+            ObjectNode.builder()
+                .withMember("name", Node.from(packageName))
+                .withMember("version", Node.from("1.0.0"))
+                .build(),
+        )
+        .withMember(
+            "build",
+            ObjectNode.builder()
+                .withMember("rootProject", true)
+                .withMember("generateServiceProject", true)
+                .withMember(
+                    "optInAnnotations",
+                    Node.arrayNode(
+                        Node.from("aws.smithy.kotlin.runtime.InternalApi"),
+                        Node.from("kotlinx.serialization.ExperimentalSerializationApi"),
+                    ),
+                )
+                .build(),
+        )
+        .withMember(
+            "serviceStub",
+            ObjectNode.builder().withMember("framework", Node.from("ktor")).build(),
+        )
+        .build()
+    val outputDir: Path = Paths.get("build", "service-constraints-test").also { Files.createDirectories(it) }
+    val manifest: FileManifest = FileManifest.create(outputDir)
+
+    val context: PluginContext = PluginContext.builder()
+        .model(defaultModel)
+        .fileManifest(manifest)
+        .settings(settings)
+        .build()
+    KotlinCodegenPlugin().execute(context)
+
+    val bearerValidation = """
+        package $packageName.auth
+
+        public fun bearerValidation(token: String): UserPrincipal? {
+            if (token == "correctToken") return UserPrincipal("Authenticated User") else return null
+        }
+    """.trimIndent()
+    manifest.writeFile("src/main/kotlin/$packagePath/auth/Validation.kt", bearerValidation)
+
+    val settingGradleKts = """
+        rootProject.name = "service-constraints-test"
         includeBuild("../../../../../")
     """.trimIndent()
     manifest.writeFile("settings.gradle.kts", settingGradleKts)
